Music logic moved into view model
