+++ app/src/main/java/com/vk/chillify/presentation/templates/navigation/BottomNavigationBar.kt
@@ -1,6 +1,5 @@
-package com.vk.chillify.presentation.templates
+package com.vk.chillify.presentation.templates.navigation
 
-import android.media.MediaPlayer
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.size
@@ -12,15 +11,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.vk.chillify.R
 import com.vk.chillify.presentation.templates.MusicPlayerBar
-import com.vk.chillify.presentation.templates.navigation.topLevelRoutes
 
 @Composable
 fun BottomNavigationBar(navController: NavController) {
@@ -34,7 +30,7 @@ fun BottomNavigationBar(navController: NavController) {
                     }
                 }
             },
-            mMediaPlayer = MediaPlayer.create(LocalContext.current, R.raw.govnovoz)
+
         )
